Fix 50-day average window and keep averages as doubles
The 50-day fetch used the 20-day start date, so both averages were
identical. averagePrice also returned int, truncating the decimals.

+++ engine/src/average_price.cpp
@@ -2,7 +2,7 @@
 #include "vector"
 #include "nlohmann/json.hpp"
 
-int averagePrice(const std::vector<nlohmann::json>& stockData){
+double averagePrice(const std::vector<nlohmann::json>& stockData){
     double total = 0;
     int count = 0;
     for (const auto& item : stockData){
