+++ engine/src/average_price.h
@@ -1,4 +1,4 @@
 #include "vector"
 #include "nlohmann/json.hpp"
 
-int averagePrice(const std::vector<nlohmann::json>& stockData);
+double averagePrice(const std::vector<nlohmann::json>& stockData);
